Moved packet_in handler to manager module.
