+++ vnet/lib/vnet/manager.rb
@@ -12,6 +12,7 @@ module Vnet
     include Vnet::Event::EventTasks
     include Vnet::Event::Notifications
     include Vnet::LookupParams
+    include Vnet::ManagerPacketHandler
 
     # Main events:
     #
@@ -116,18 +117,6 @@ module Vnet
     #
     # Other:
     #
-
-    def packet_in(message)
-      if (message.cookie & COOKIE_DYNAMIC_LOAD_MASK) == COOKIE_DYNAMIC_LOAD_MASK
-        handle_dynamic_load(id: message.match.metadata & METADATA_VALUE_MASK,
-                            message: message)
-      else
-        item = @items[message.cookie & COOKIE_ID_MASK]
-        item.packet_in(message) if item
-      end
-
-      nil
-    end
 
     def set_datapath_info(datapath_info)
       if @datapath_info
@@ -607,49 +596,5 @@ module Vnet
       }
     end
 
-    #
-    # Packet handling:
-    #
-
-    # TODO: Move to a module.
-
-    def handle_dynamic_load(params)
-      item_id = params[:id]
-
-      debug log_format('handle dynamic load of item', "id: #{item_id}")
-
-      return if !push_message(item_id, params[:message])
-
-      item = internal_retrieve(id: item_id)
-      return if item.nil?
-
-      return item
-    end
-
-    # Returns true if the message queue was empty for 'item_id'.
-    def push_message(item_id, message)
-      return if item_id.nil? || item_id <= 0
-      return if message.nil?
-
-      # Check if the item got loaded already. Currently we just drop
-      # the packets to avoid packets being reflected back to the
-      # controller.  
-      return if @items[item_id]
-
-      if @messages.has_key? item_id
-        # TODO: Cull the message queue if above a certain size.
-        @messages[item_id] << {
-          :message => message,
-          :timestamp => Time.now
-        }
-
-        return false
-      end
-
-      @messages[item_id] = [{ :message => message,
-                              :timestamp => Time.now
-                            }]
-      true
-    end
   end
 end
