+++ vnet/lib/vnet/manager_packet_handler.rb
@@ -1,0 +1,67 @@
+# -*- coding: utf-8 -*-
+
+module Vnet
+
+  module ManagerPacketHandler
+
+    def packet_in(message)
+      if (message.cookie & COOKIE_DYNAMIC_LOAD_MASK) == COOKIE_DYNAMIC_LOAD_MASK
+        handle_dynamic_load(id: message.match.metadata & METADATA_VALUE_MASK,
+                            message: message)
+      else
+        item = @items[message.cookie & COOKIE_ID_MASK]
+        item.packet_in(message) if item
+      end
+
+      nil
+    end
+
+    #
+    # Internal methods:
+    #
+
+    private
+
+    def handle_dynamic_load(params)
+      item_id = params[:id]
+
+      debug log_format('handle dynamic load of item', "id: #{item_id}")
+
+      return if !push_message(item_id, params[:message])
+
+      item = internal_retrieve(id: item_id)
+      return if item.nil?
+
+      return item
+    end
+
+    # Returns true if the message queue was empty for 'item_id'.
+    def push_message(item_id, message)
+      return if item_id.nil? || item_id <= 0
+      return if message.nil?
+
+      # Check if the item got loaded already. Currently we just drop
+      # the packets to avoid packets being reflected back to the
+      # controller.  
+      return if @items[item_id]
+
+      if @messages.has_key? item_id
+        # TODO: Cull the message queue if above a certain size.
+        @messages[item_id] << {
+          :message => message,
+          :timestamp => Time.now
+        }
+
+        return false
+      end
+
+      @messages[item_id] = [{ :message => message,
+                              :timestamp => Time.now
+                            }]
+      true
+    end    
+
+  end
+
+end
+
